updated layout of the project so that debugging works
