+++ .gitignore
@@ -1,3 +1,3 @@
 **/node_modules/*
-CherryIdentities/db.sqlite3
+*.sqlite3
 **/package-lock.json
