--- CherryIdentities/src/ormconfig.ts
@@ -1,15 +1,0 @@
-export default {
-    type: 'sqlite',
-    database: './db.sqlite3',
-    logging: ['error', 'warn', 'schema'],
-    entities: ['node_modules/@jolocom/sdk-storage-typeorm/js/src/entities/*.js'],
-    /** or if you list entity classes, then simply add the SDK entities
-    entities: [
-      // your entities here
-      // then
-      ...require('@jolocom/sdk-storage-typeorm').entityList
-    ],
-    */
-
-    synchronize: true,
-}
