--- CherryIdentities/text.md
@@ -1,3 +1,0 @@
-# Creating Identies with Jolocom-SDK
-
-The necessary node modules are not uploaded
