--- Interface/styles.css
@@ -1,5 +1,0 @@
-﻿body {
-    background-color: #F0F0F0;
-    padding: 2em;
-    font-family: 'Raleway','Source Sans Pro', 'Arial';
-}
