+++ src/SmartContractInterface/Interface/.vs/ProjectSettings.json
