+++ src/SmartContractInterface/Interface/.vs/slnx.sqlite
