+++ src/SmartContractInterface/Interface/script-onlyClient.js
