+++ src/WebInterface/styles.css
@@ -1,0 +1,5 @@
+﻿body {
+    background-color: #F0F0F0;
+    padding: 2em;
+    font-family: 'Raleway','Source Sans Pro', 'Arial';
+}
